+++ app/src/main/java/de/example/crudapp/repository/ProductRepository.kt
@@ -14,7 +14,7 @@ interface ProductRepository {
     suspend fun createProduct(
         name: String,
         description: String,
-        price: Float,
+        price: Double,
         qty: Int
     ): Response<Product>
